Split price elasticity in I and O
Supply and demands have now two products, one for the "input" elements, for which we use the composite demand prices, and
one for the "output" elements, for which we use the composite supply prices.

The supply of a product will then depend on the (demand) prices of the input factors (if any) and on the product (supply) price,
while the demand will depend on the (demand) price of the product and possible substitute/complements and the (supply) price of
the possible transformed products (if any).

Note that composite demand and supply prices for a product/region are not necessarily equal as these are aggregate of
heterogeneous products.